requirements.txt
asgiref==3.5.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.0
Django==4.1.1
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
idna==3.4
importlib-metadata==5.2.0
isodate==0.6.1
ldap3==2.9.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
platformdirs==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
python-decouple==3.6
python-ldap==3.4.0
pytz==2022.7
requests==2.28.2
requests-file==1.5.1
requests-toolbelt==0.10.1
six==1.16.0
sqlparse==0.4.3
suds==1.1.2
urllib3==1.26.15
Werkzeug==2.2.2
zeep==4.2.1
zipp==3.11.0
